requirements.txt
pyTelegramBotAPI==4.32.0
python-docx==1.2.0
pdfplumber==0.11.9
google-genai
